[java client] Cleanup TestKuduSession and KuduClient
The second test in TestKuduSession is using AUTO_FLUSH_SYNC but it's
only joining on the last Deferred, leaving it open to out-of-order acks.
There are a few other places that do weird things with Deferred, those
are also cleaned up. Some joins() were missing a timeout and now use
DEFAULT_SLEEP.

KuduClient had an unused property from the previous commit.

Change-Id: Ibb14377f32c0735f94ef55179eafe6282263ec08
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3423
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>