knapsack_solver-test: use ASSERT_NEAR instead of ASSERT_DOUBLE_EQ
Apparently the ASSERT_DOUBLE_EQ macro doesn't have a large enough epsilon
to deal with the errors that accumulate inside the knapsack algorithm.
We saw the following test failure:

  /var/lib/jenkins/workspace/kudu-test/BUILD_TYPE/LEAKCHECK/label/centos6-kudu/src/util/knapsack_solver-test.cc:126: Failure
  Value of: sum_val
    Actual: 872.43508785666847
  Expected: max_val
  Which is: 872.43508785666904

which is les than a 1e-12 difference... but still failed the test.
Now we use ASSERT_NEAR with a much bigger epsilon to avoid false positive
errors like this.

Change-Id: I6307659ce957feb9e67d049b483f709c52a3a0a9
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3430
Tested-by: jenkins
Reviewed-by: Vladimir Feinberg <vladimir.feinberg@cloudera.com>