tablet: refactor some test base code into a new TabletHarness class
In the spirit of "implementation inheritance is a mess", refactoring some
harness code for testing tablets into a new class. I've found that sharing
utility code for testing tablets via the base class is somewhat messy.
Refactoring out the tablet with its dependencies into a harness class
is a little easier to follow.

This also does a small amount of non-test-code change - in particular
the TabletPeer no longer maintains its own anchor registry, but instead
just uses the already-existing pointer held by the tablet itself. This
simplified the wiring a tiny bit.

I plan to follow this up with a few other things:
- create an equivalent TabletPeerHarness
- try to move the InsertForTesting code into TabletHarness
- eventually refactor the tests that currently run directly against
  TabletHarness to run against a TabletPeer instead

Change-Id: Iff035f3b520c9bbd57ab3b23e6d086e5da4b74f1
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3431
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>