Avoid unnecessary unlock on a guard
Unlocking a guard is annoying unless absolutely needed, in this case
scoping takes care of the issue with the possible "delete this"
scenario. We also bring the initial CHECK inside the critical section,
since it makes crashes easier to understand.

Also add comment noting that unregistering maintenance ops is not thread
safe.

Change-Id: Idecfee1bd033f8260f808ddf8cddaf1c393d52d3
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3451
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>