KUDU-392. Expanded Random Interface.
This commit adds the convenience method Next64() to the random class in
util/random.h. Additionally, the many parts of the code where equivalent
behavior was simulated were replaced by uses of Next64().

Additionally, in util/random_util.* a similar method was added which allowed
for arbitrary-size byte strings.

Only a few, easy-to-change tests were replaced with the new Random interface.
Eventually, all calls to rand() should be replaced with a Random instance
to promote centralization and deduplication. This would perhaps be possible
by having a Random instance in KuduTest.

In the future, please rely on Random instances, initialized with the return
value of SeedRandom(), rather than rand.

Change-Id: Ieed6fc91054b227bcab9ff2bdc630d329ee5fefc
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3443
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins