Fix HighWaterMark::IncrementBy(), add unit test.
Correct behavior for HighWaterMark::IncrementBy() is to increment
current value by delta and if the result (current value + delta) is
greater than the current max, update max value. Initial revision would
only update the max value to the delta and not the result, resulting in
incorrect peak consumption for MemTracker.

This revision fixes the issue and adds a unit test for HighWaterMark.

Change-Id: I9eb7ea283709d83c7e45045a7cbbd8725c680a79
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3448
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins