Move operation decoding and rowlock acquisition from WriteTransaction into Tablet
This moves most of the "Prepare" phase code from WriteTransaction into the
Tablet class itself

The code for preparing a Write is currently somewhat duplicated in three places:
  1) in WriteTransaction (moved in this patch)
  2) in TabletBootstrap (untouched by this patch)
  3) in Tablet::*ForTesting methods (untouched by this patch)

This is messy due to code duplication, and also messy in that the Prepare
phase for a write involves some tricky coordination with tablet-internal
locking: in particular it needs to acquire row locks from the tablet's
lock manager, and soon will also need to perform some locking around
the tablet's schema (to fix the alter_table-test races plaguing us)

So, this patch does a small bit of cleanup by moving the prepare code from
WriteTransaction into Tablet itself. This patch does not yet address
consolidating code with the other two areas mentioned above. But, with the
"main implementation" for preparing a write now living within Tablet, it becomes
somewhat easier to consolidate the other code in the next patch in this series.

In order to do this code move, the WriteTransactionState now contains a
vector<DecodedRowOperation>, rather than the ops being in local function scope
within WriteTransaction::CreatePreparedInsertsAndMutates. This may have a small
negative performance impact in the short term, but if it's problematic, it
should be straightforward to make these allocations come from the transaction's
Arena.

Change-Id: I50d51c4cc5f5b44ca006acc8b157ae6437f3bed8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3765
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>