KUDU-46: PIMPL KuduSession
After KuduScanner comes KuduSession. Here is where Data being public helps
us: it means Batcher and KuduClient can access members of Data without us
having to expose those members in KuduSession and by extension, to users.
However, these classes still need to be declared friends of KuduSession, so
that they can "unwrap" a session and access the tasty Data object inside it.

One other wrinkle: because KuduSession is a shared object but
KuduSession::Data isn't, Data functions that need to access KuduSession
are given a shared_ptr to it. This is in lieu of a dedicated back pointer
and is the recommendation from http://herbsutter.com/gotw/_100. Affected
functions are:
- Init
- NewBatcher
- Apply

Change-Id: I15d74d59f3143afbce91f95002f43b4df33a1d16
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3509
Tested-by: jenkins
Reviewed-by: Vladimir Feinberg <vladimir.feinberg@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>