tests: clean up test files in KuduTest destructor instead of TearDown
This changes the removal of the test directory to happen in the destructor
of the test rather than in the TearDown method. This ensures that the user's
destructor runs while the test directory is still present.

In particular, this fixes an issue with linked_list-test where the test directory
would get removed before the minicluster was shut down. This cased CHECK failures
in the subprocesses which made test logs confusing to read and occasionally
triggered spurious TSAN failures.

Change-Id: Ibd9dbbdf4b97ab1b3e4b78c209ca34e809153114
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3470
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>