Added new default values for flags in full_stack-insert-scan-test
Now, the default values depend on whether AllowSlowTests() from KuduTest
returns true or not. If false, 1500 rows are inserted (not enough to cause
a flush - intentional, to keep unit tests fast). If true, then 500K rows
are inserted, which does cause both flushes and compactions to occur in
the WithDiskStressTest.

Manually specifying flags overrides both defaults.

Change-Id: I1e29f3c784526cd0eae91fc1473623d353f3b79c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3474
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>