KUDU-46: PIMPL'ed builder for KuduClient
KuduClientOptions isn't a true builder, so let's add one. It should be
PIMPL'ed, as we will no doubt add more options in the future.

One point of frustration: because it's PIMPL'ed, I don't think it makes
sense to allow copying/assigning. But that means we have to modify the
client's builder directly in ExternalMiniCluster.

Change-Id: Ia7c4ad068148b0cf63a6ed3f846e5e5ee1b398c1
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3515
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>