KUDU-46: PIMPL'ed builder for CreateTable
We should PIMPL CreateTableOptions, because they're likely to change in the
future and a compilation firewall will give us some semblance of backwards
compatibility.

While I was there, I converted it into a true builder, which also meant
moving the CreateTable verb itself from the client to the builder.
Admittedly this last part is a stylistic choice, but a quick poll on Hipchat
confirmed that others felt the way I did.

Change-Id: Ib05a7392ed54078979ae4b7b3121d74bf9b0bb48
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3516
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>