get rid of boost filesystem/system usage
Looks like we might have to bring boost into thirdparty, so let's do what
we can to minimize the amount of boost we use. Getting rid of filesystem is
trivial, but despite getting rid of local references to system, thread
depends on it. So it's not going anywhere until thread is gone.

Change-Id: I19fce36f5d84ed928f2ff0ec5e76c8fdf826dbff
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3521
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins