KUDU-230, KUDU-63. Import/implement our own Mutex and ConditionVariable
This imports the condition variable and mutex from Chromium, with a certain
amount of changes to fit into our source tree.

I also included wrapper functions for the boost naming convention so that
boost::lock_guard can still be used with the new mutex. We'll remove these
over time.

I converted a couple of existing boost::mutexes over in this commit. The rest
will follow later.

Change-Id: I626b7144ae8bf02f884e417a881e8b736459bc86
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3522
Tested-by: jenkins
Reviewed-by: Vladimir Feinberg <vladimir.feinberg@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>