thread: replace boost thread with direct pthread
To make the client PIMPLing work, we need to either rid ourselves of the
dependencies on the boost thread and system libraries, or move them into
thirdparty and rebuild them with -fPIC. I was leaning towards the latter,
but Todd has recently taken it upon himself to rid us of boost::mutex and
friends, making the former more tractable.

To that end, here's kudu::Thread implemented directly on top of pthreads.
What are the notable changes?
- We no longer need to use ScopedThread; we can use pthread_cleanup_push()
  and pthread_cleanup_pop() to handle cleanup. Besides less code, this also
  gets pthread_exit() right. Hat tip to Colin for the suggestion.
- pthread_timedjoin_np() is non-portable and operates in absolute wallclock
  time, which can be affected by system clock changes. Instead, we'll use a
  CountDownLatch for timed joining. The one caveat is that we still need to
  call pthread_join() when we're done, otherwise we're not guaranteed to
  have pthread-driven cleanup (like pthread_key_create() destructors) called
  by the time we finish joining.
- Passing data from parent StartThread() to child SuperviseThread() is a
  little harder. To that end, I've simplified the parent<-->child handshake
  so that it uses just tid_.

Change-Id: Ib2b248bcde5392453d22cdab9b9410e494ad573c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3677
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>