KUDU-46: replace boost lock_types in code consumed by client
The public client headers don't contain any references to boost, however the
various client innards do, which means that libclient.so needs to link
against some boost libraries. It can only do this if the host system has
boost static archives compiled with -fPIC, which very few hosts do.

Instead of compiling boost with -fPIC in thirdparty, we can get rid of the
various references that still exist: libboost_thread and libboost_system.
The former is gone when we replace the boost::thread usage in util/thread
with direct pthreads. The latter is gone when we remove all outstanding
references to boost::lock_guard, boost::unique_lock, and boost::shared_lock.
With new versions of boost, the classes in these methods throw, which means
they need libboost_system.

Change-Id: I878b28de2d9caf2b258743b5805fae20fa543ec0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3678
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>