KUDU-131 - Part 1 - Separate transaction Start() from Prepare() and Apply()
Right now we're setting transaction timestamps on Prepare(). Once
we change the FOLLOWER/REPLICA transaction implementation to accept the
LEADER's timestamp we need to do it right before Apply() vs. on Prepare() as
we don't have the timestamp yet then. Moreover, as was introduced in the design doc,
we need to have it happen in the thread that is triggering the applies (vs. on
the Apply() thread itself), otherwise we will have races in safe time
advancement.

This adds a new method to Transaction implementations, Start(), that does this.
Leader/ReplicaTransactionDriver then make sure to call it at the right times.
Note that actually accepting the LEADER timestmaps will come in a follow up patch.

Change-Id: I6e2091f19bed6c6202ea8ebe1454d09ee8f7b42b
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3537
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>