Documented non-empty rowblock requirement to ConvertRowBlockToPB()
Previously, if the RowBlock supplied to ConvertRowBlockToPB was of size 0
(note that this is not the same as a nonzero size with a clear selection
vector), then a DCHECK triggered in the CopyColumn() method where one of the
column block's cell_ptr is accessed. That method has a check to make sure that
the indexed accessed is less than the size. In the case that the size is 0, this would cause an error.

Rather than inserting a check for this at the lower level, it is more
appropriate to document the edge case and let the caller know that the
RowBlock should not be empty, since it does not make sence to send a
size-0 rowblock.

Change-Id: I89b242a94b2fa475b15a13018f3bd3a60714b219
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3542
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>