CBTree depth reduction
Removed size from the cbtree valueslice class (placing it in the pointed-to
data instead). This allows us to fit more data into a leaf, which reduces tree
depth in a few rare cases.

At the very least, it doesn't slow down scanning, since scans which require
the length of a slice dereference the slice's data pointer anyways, so the
cache miss is already present.

This commit also gets rid of a few overconstrained traits that the cbtree has,
allowing the user to only specify the desired size of the internal and leaf
nodes, and maximizing the amount of entries that can fit in there (technically,
the traits weren't overconstrained, because the user could reduce the number
of entries present in order to increase the inline key size, but all this
really did was create an additional dependency on the user to not be
inconsistent in requesting too many entries while keeping the size of the
nodes the same). Since the inline key size almost always worked out to be
8 bytes, and usually this is enough anyway, I removed this option from being
configurable.

Before:
 Time spent concurrent inserts (10000000 rows, 50 threads): real 12.001s
 Time spent empty projection, 0 col: real 0.359s
 Time spent key scan, 1 col: real 0.910s
 Time spent full schema scan, 10 col: real 5.245s
 Time spent String projection, 1 col: real 0.883s
 Time spent Int32 projection, 4 col: real 1.506s
 Time spent Int64 projection, 4 col: real 2.230s

After:
 Time spent concurrent inserts (10000000 rows, 50 threads): real 11.660s
 Time spent empty projection, 0 col: real 0.355s
 Time spent key scan, 1 col: real 0.888s
 Time spent full schema scan, 10 col: real 5.156s
 Time spent String projection, 1 col: real 0.872s
 Time spent Int32 projection, 4 col: real 1.619s
 Time spent Int64 projection, 4 col: real 1.861s

Change-Id: I7521db2370a46b3ed9d86374bdc093cbb1f98d5b
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3564
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>