KUDU-46: replace gtest.h with gtest_prod.h
Many headers use FRIEND_TEST and include all of gtest to get it when there's
a simpler alternative: gtest_prod.h. This is especially important for the
exported client library, where shipping all gtest headers is painful.

I built this list by looking for FRIEND_TEST definitions in headers and
ignoring files with "test" in the name.

Change-Id: I8960767843db9030f9392b70852e779dd2c3a4f3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3679
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>