KUDU-335: Share apply executors across tablets.
This revision moves initialization of
executor for leader and replica apply tasks from TabletPeer
to its owner (SysTables in Master and TsTabletManager
in Tablet).

The apply executors are now owned, initialized, and shutdown by
SysTables and TsTabletManager.

This doesn't change behavior for the leader
apply executor (which was already multi-threaded), but changes
replica apply executor to be multi-threaded: this was an
existing TODO item and should be safe.

Additionally, this revision adds a state enum and asserts
to TsTabletManager to make sure that new TabletPeer instances
can't be created once the apply executors have been shutdown.

TODO If feasible, do the same for log GC and prepare executors.

Change-Id: If9646b2caa1c216cc01c9eac1ff9bd363ddfcb57
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3599
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins