KUDU-405 Part 1. Initial codegen commit.
This commit creates the basic codegen classes - a central code_generator
that synchronously generates classes which rely on codegen'd functions, one
such class, a row projector, and the FunctionBuilder, which for now
is just a convenience wrapper around the llvm::IRBuilder<> class (necessary
to keep llvm headers out of non-source-file includes.

Currently, there is no code cache or asynchronous dedicated compilation thread.
These features are to come in future commits. Thus as of now the code
always compiles and waits for compilation to finish during a query.

Furthermore, the code projector is not yet integrated into the MRS.

The row projector has the same functionality as the non-code-generated one,
but only for the projection ContiguousRow -> RowBlockRow.

Finally, this commit turns off the codegen test in the CLIENT
jenkins build. This is because certain symbols, such as Arena's member
methods, are hidden in that build. Since the codegen component requires
runtime symbol resolution, it fails in this mode.

Change-Id: I15863babcbf3ec984641831f038dda0b76fab1d5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3704
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins