thirdparty: more thorough handling for llvm-building-llvm nonsense
When I added llvm to thirdparty, I hard-coded the "use llvm 3.3 from the
toolchain on RHEL6 to build llvm 3.4" logic in build-and-test.sh.
Unfortunately, this broke the tpch1 and benchmarks builds, which also run on
RHEL6 and build thirdparty.

Here's a more comprehensive fix. The logic for using llvm 3.3 moves to
build-thirdparty.sh and is generalized somewhat: if CC and CXX aren't set,
and if we can find clang on the PATH, we'll use it to build llvm 3.4.
Otherwise, we'll go with the system compiler. Now the benchmarks and tpch1
builds need only source the toolchain script (to add llvm 3.3 to the PATH)
and they should be handled properly.

As part of this commit, I moved the bulk of the tpch1 logic from the Sandbox
job and into a new script.

Change-Id: I43f9a42e59d8279fe4c926583c3a7020cdea520c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3667
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>