cbtree: remove Update() function
Per KUDU-407, there were some bugs in the Update path. When I went to fix them,
I realized that we never in fact use Update() anyway, and the support for it
was a little half-baked. So, in the spirit of reducing code footprint, I just
removed the Update ability from the concurerent btree.

Change-Id: I5eb1369508cbd52892b2b49c94d58caaf2a34efb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3673
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>
Tested-by: jenkins