tpch: better result parsing
In the first tpch run that actually succeeded after I broke everything, I
saw this:

  src/scripts/write-jobs-stats-to-mysql.py:27: Warning: Data truncated for \
    column 'runtime' at row 1 (job_name, build_number, workload, \
    iteration, runtime))

And in the database there was a 0 for the runtime of that row.

Turns out it's because the existing parsing failed when the logged tid
wasn't 5 digits long: the number of delimiters in the line increased and
we ended up with "real" as the value for runtime.

In lieu of robust JSON-based result dumping and parsing, let's at least fix
this with some perl regexes.

Change-Id: I412f8794a5c6427237cfe377c1af283f620c6fb7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3674
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>