Define utility functions in test_util.h and test_util.cc
Add a kudu_test_util library to hold these functions, which are factored
out of the KuduTest class. This allows use of these functions without
having to plumb a reference to KuduTest to the caller.

Also tweak a few usages of AllowSlowTests() to not specifically
dereference 'this' to call the functions.

Change-Id: I71b547a55412663356d22f879e85cdb6e043d88b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3694
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins