KUDU-46: rename a bunch of public client stuff to include kudu
When the public client is installed, it should make clear reference to
"kudu". Right now it doesn't. For example, the library is called
libclient.so. So let's fix that.

Here is a comprehensive list of renames:
- libclient.so -> libkudu_client.so
- $LIBDIR/client/cmake --> $LIBDIR/kuduClient/cmake
- $DATADIR/samples --> $DATADIR/kuduSamples
- cmake/client*.cmake --> cmake/kuduClient*.cmake

There's an unfortunate discrepancy between camelCase and under_scores that's
due to how cmake's find_package() works: if I call find_package(foo), it
looks for fooConfig.cmake. Hence the (less desirable) camelCase.

Change-Id: I66ce98262377767036be382f2bea9cd3877af9c4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3701
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>