Initial commit of centralized test failure reporting
This adds a simple Python web server which takes POST requests
recording test results. The results are inserted into a MySQL table
which records the revision, test name, exit status, etc. Test
logs are also uploaded to S3 with a UUID key, and the UUID is
recorded in the database.

The run-test.sh script is also updated to call a new report-test.sh
script which actually sends the results to the HTTP server.

I tested this manually locally and verified that records showed up
in the SQL database:

  *************************** 264. row ***************************
            id: 264
     timestamp: 2014-08-01 18:31:35
      build_id: non-jenkins
      revision: 6149228e32bfe3c8071c3deeb1215514fdf07961-dirty
  build_config: fastdebug asan
      hostname: todd-ThinkPad-T540p
     test_name: oauth-test
        status: 23
       log_key: bdba2b50-19e4-11e4-a54d-005056c00001

In a later commit, I'll add some simple web interface to track basic
statistics - in particular, for flaky tests, I want to see a percentage
"pass rate" for recent revisions, so it's easier to see where they got worse.

Change-Id: I7f75a6b04190dcc2b8fd4e407f5e1837d0e89e32
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3732
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins