KUDU-46: unit test for checking symbols in the exported client library
This new test checks that all the symbols in the exported client library
are acceptable. Any symbol in the dynamic symbol table that is both global
and strong (i.e. a client can resolve it) must be in the kudu namespace.

Oh, and now ADD_KUDU_TEST supports any script-like unit test.

Change-Id: I8aef9e740c3bb980c49cf0c4cd29350f0a540da3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3760
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>