KUDU-296: mutate some rows as part of linked_list-test
The test now kicks off a background thread that receives flushed keys via a
blocking queue. For each one, it updates the corresponding row and sets the
new 'updated' column to true. When we're done inserting, we kill the thread.
During verification, we make sure that every row's 'updated' column is true.

Due to KUDU-333, mutations are off by default.

Turns out we didn't have BOOL support in KuduRowResult or KuduPartialRow, so
I added that too.

Change-Id: I1882cc7c8de5eb9523cee73592ad656f06a35f59
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3777
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>