client: common RPC layer
Well, it's a bit of a stretch to call it a "layer", but so it goes. The
point is to consolidate the retry logic that's common to both writes and
lookups in one place, which I think this class hierarchy manages to do.

The RPC itself is merely an interface; the retry logic is in RpcRetrier.
There's good encapsulation here, though it means slightly more complexity
in each instance's SendRpcCb.

Change-Id: I03ed4353f562eeb9e27c7f202a495a22ccea2978
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3818
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>