Change alter_table-test to use client APIs
This changes alter_table-test to use the client APIs to create and
alter the table. Since the test is in the integration-tests directory
it makes more sense to use public APIs rather than directly accessing
the master's internal functions, etc.

This also adds timeout support to the public alter table function.

Change-Id: I72d5cc4aed16705734348f105a2604f822277708
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3811
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>