catalog_manager: prevent alter table from completing too early
In multi-tablet tables, the existing code could finish an alter table before
every tablet was altered. This commit fixes that and introduces a unit test
that failed most of the time before the fix.

I also added a second test that alters a table a bunch of times, waits on
the alters to finish, and tests the schema. Unfortunately it breaks because
of other alter table bugs. For now it's disabled, as it brings a few other
features along for the ride, such as KuduTableAlterer::wait() and
KuduSchema::num_cols().

Oh, and apparently this fixes KUDU-273. What fun!

Change-Id: I7faafec939f98da049451c59c948ec993f4fe0b6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3833
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>