ensure there's no test data left behind when a test passes
The way that MiniCluster and ExternalMiniCluster tried to reuse test
directories wasn't working properly: every call to GetTestDirectory creates
a new directory because the time of day is appended to the directory name.
Todd suggested storing the time of day in a static variable so that
GetTestDirectory returns the same thing every time in a test. That's not too
magical, and it works, so it's what I've done.

Oh, and I modified build-and-test.sh to test for this condition. More logic
for Todd to port!

Change-Id: I3513c6e5b952e3aae87d1b760621d04dd0464e89
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3848
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>