KUDU-405 Part 2. Added type safety to precompiled functions
In order to gain some type safety with the LLVM method calls,
a new context has to be created per module. This actually does not
incur an additional memory overhead because the contexts are deleted
after compilation. The new contexts allow us to query the parsed
LLVM module for the type required by name, directly.

Also made some other minor changes:
1. Added module verification
2. Removed unneeded headers
3. Changed GoogleOnceType usage in code_generator.cc.
4. Increase ModuleBuilder encapsulation (remove module() getter)

Because of the new ownership semantics, ModuleBuilder now owns the
LLVMContext and Module which get deleted after compilation.

Change-Id: I77d4538c38e8b1c01bee9cf75a610812f81c49fb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3867
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>