KUDU-465 - Part 1 - Only replicate after prepare has finished
With the necessity for the propagation of timestamps from the leader to replicas
we need to make the timestamp a first-class element of the state machine. Currently
it isn't as its wrapping message (the commit message) is not majority replicated,
raising state machine recovery issues. So the shortest path here is to make prepare
acquire the timestamp and only then replicate the request (as the timestamp can
only be acquired after the locks have been obtained).

This patch doesn't actually make the timestamp part of the ReplicateMsg
just changes the wiring so that consensus::Replicate() only happens after we've
acquired the timestamp on the end of Prepare().

Change-Id: Ib685e230d794897b689e2167f9b620cbd0de7473
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3888
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>