KUDU-405 Part 4. Add error checking to row projector relocations.
If a projection requires relocations (as is the case with STRING column-
containing schemas), then it admits the possibility that there is not enough
space in the relocation arena to copy the string over.

Previously, the codegen projector would just abort in this situation.
Now, it returns a error message.

Change-Id: I628917d6c3137fd92d58424ac8645e2e0203a589
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3920
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Vladimir Feinberg <vladimir.feinberg@cloudera.com>