Fix bug where UNDO deltas would not restore 'NULL' values
This fixes an issue in ColumnBlock::Cell where we ignored the nullability
of the cell in the .ptr() method. This had the downstream effect that,
when generating UNDO deltas for updates which transitioned a cell from
a NULL to non-NULL value, we created a garbage UNDO.

This wasn't caught by existing tests because we didn't check snapshot
consistency for nullable columns which flip between NULL and non-NULL.
Neither did we specifically test for the UNDO generation in this case.
I amended compaction-test to do the latter.

Change-Id: Ic45c61451c4dd9efb042924369b9784a1f0e9f12
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3906
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>