Fix unintentional memory sanitizer includes in clang headers.
When thirdparty's clang is built with a local clang instance,
it is built a configuration file which #defines macros that indicates
that sanitizer/msan_interface.h is available to the compiler.

When rebuilding kudu with gcc, gcc does not have this include
available, so it fails compilation. This commit resolves the problem
by creating a header that should be included before any llvm include.
It #undefs the appropriate macro so that builds are compatible with
compilers other than clang.

Change-Id: I0e23a9f45a76844cff38884d1c1b02a93be0e47b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3915
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>