KUDU-412 Some parts of CreateTablet should be aysnc
This repurposes the bootstrap pool in the tablet manager to open all the tablets, not just
those that exists when the TS starts. This fixes our issue where the CreateTablet calls are
deadlocking on themselves, but unfortunately we still have an issue with followers waiting
to be created while leaders are trying to get a quorum. This means we still have a deadlock
but at a lower level.

The most random change in this patch in where we stop marking a tablet dirty when registering
them. The problem lies in ts_tablet_manager-test, which expects that after creating a tablet
that it will have only one dirty entry for it. The problem is that we mark dirty:
 - When registering
 - When opening

The former is done sync with the CreateTablet call while the latter is now done async. Since
there's only one entry in the dirty tablets per tablet, we don't know if it means that we saw
both the registration and the opening when we see a dirty tablet, or if it just registered but
is about to open.

Since marking dirty for both seemed redundant, this patch removed the former case for marking as
dirty.

Another change is that the pool that open tablets is now as big as the number of rpc handlers. This
is so that we regress on how many tablets we can create in a table.

Change-Id: I992532fea169c3b42f412adceca6c53990d3a1b8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3930
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins