KUDU-405 Part 3. Codegen integration into MRS.
In order to start using the row projector in the MRS, several things needed
to occur:

1. The compiled row projector needed to keep track of the mappings generated
by the original row projector because they are used by different components
within the MRS iterator (hopefully, there will be codegen adaptations for
these components eventually, eliminating the need for a wrapper).
2. A tablet-server-wide compilation manager needs to be maintained. Currently,
it does not do anything except pass on a synchronous command to the code
generator, but eventually the manager will manage the code cache.
3. Now that code generation is used somewhere other than codegen/, it needs
to be disabled in KUDU_EXPORTED_CLIENT builds because hidden symbols
prevent its operation. A compile-time disabling mechanism has been added.

The manager is process-wide because using the largest possible scope
for code management will minimize redundancy and increase cache hits (once a
cache is there). This may result in a slight discrepency between real-life and
a testing enviornment which has multiple tablet servers in one process,
but the advantages of a singleton CompilationManager outweigh the losses
(see codegen/compilation_manager.h).

Change-Id: Ied9cc07d14d32bf4e4295a1dc7686337b56a87f7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3932
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>