Make Log::Reserve() take a LogEntryBatchPB instead of the OperationPBs
This adds a helper method to do what previouly Log::Reserve() did and makes
Log::Reserve() do what previously Log::DoReserve() did, i.e. actually appending
the entry batch to the log.

The reasoning is that we shouldn't be constrained to just adding  OperationPBs to the
LogEntryBatchPB. In particular as we move away from having replicated commit
messages those will no longer have ids and thus shouldn't be a part on an OperationPB,
i.e. and should be added to a batch by themselves.

Change-Id: Ica387bd334a60210d21ca26b8a23392886c7f9c5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3933
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>