arena: avoid an Acquire barrier in non-threadsafe arena
The single-threaded Arena implementation was using an unnececessary
Acquire_Load to load the current Arena::Component. This turns into a normal
load on x86, so it isn't a ton of overhead, but it does necessitate a compiler
barrier, which prevents any elimination of redundant loads in a loop, etc.

I don't know that this will make a noticeable difference on any benchmarks, but
I noticed the compiler barrier in some llvm IR, and tracked it down to this
spot. Fixing it might open up a few loop optimizations in our row-copying
paths, etc.

Change-Id: I5c40b74d4c646f79e71b0876c7606ce65b6ffbee
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/3970
Reviewed-by: Vladimir Feinberg <vladimir.feinberg@cloudera.com>
Tested-by: jenkins