No replicated commit messages
This buids on the log patch to erradicate commits from the replicated log
completely. Commit messages no longer have OpIds, and are only written to the
local log. The replication of commits is managed per the Raft paper -- the
leader keeps track of the replication watermark, and once a majority of
replicas have replicated a given operation, it can update its commitIndex and
propagate the new value to followers.

The main changes in this patch are:

- Commits are no longer a second class part (or any part at all) of the state
  machine. They are purely local messages.

- Bootstrap is updated to properly keep track of the latest seen commit
  message, and pass it in ConsensusBootstrapInfo

- The follower-side code is updated to use the commit index to
  mark its local operations as Consensus-committed. It no longer provides the
  leader's commit operation, since that isn't replicated.

One side effect of the patch is that we don't aggressively replicate the commit
watermark. In other words, if we were to do a single operation against a
leader, we could see the following in the worst case (with a heartbeat time
of 1 second)

  1) +0ms   Operation arrives at leader. Leader adds to consensus queue and
            triggers UpdateConsensus() calls on the followers.
  2) +5ms   Followers receive the new operation, write it to their logs, and
            respond that they have replicated it.
  3) +5ms   Leader updates its commitIndex and marks the operation as committed.
            Followers still don't know whether or not it was committed.
  4) +990ms Leader finally heartbeats, telling the followers that the operation
            is committed.

Prior to the patch, step "4" here came only a few milliseconds after step "3",
since the addition of the CommitMsg in the consensus queue triggered a new
UpdateReplicas() round. Now, we don't do that signalling, so the replicas have
to wait some time to find out about the commit.

Because of the above, this patch updates a number of unit tests to sleep for
1.5 seconds after doing some operations. This gives time for a consensus
heartbeat to trigger and pass along the new commit index to the replicas.

Todd consulted with the CockroachDB folks, and they're of the opinion that it
makes sense to eagerly trigger an UpdateConsensus() call whenever commitIndex
changes. That probably makes sense, but I didn't want to make this patch even
more complex, so I just updated the tests with workarounds here.

This patch was co-contributed by David and Todd.

Change-Id: I6e580af74c0536aaa227d7fb57a08951fb85c55c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4460
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins